% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_cases_cvera_read_in.R
\name{all_cases_cvera_read_in}
\alias{all_cases_cvera_read_in}
\title{Read the Latest bTB all_cases dataset (animal level bTB dataset)}
\usage{
all_cases_cvera_read_in(base_path = "N:/data/tb/master_tb_data_jamie_m")
}
\arguments{
\item{base_path}{Character. Path containing folders of the form
\code{year_XXXX_qX}.}
}
\value{
A dataset read from the most recent \code{all_cases*.rds} file.
}
\description{
This function finds the most recent \code{year_XXXX_qX} folder inside a base
directory, identifies \code{.rds} files whose filename contains \code{"all_cases"},
and loads the most recently modified file.
}
\examples{
\dontrun{
all_cases_cvera_read_in("N:/data/tb/master_tb_data_jamie_m")
}

}
