% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_cases_per_year.R
\name{all_cases_per_year}
\alias{all_cases_per_year}
\title{all_cases_per_year}
\usage{
all_cases_per_year(
  df,
  drop_years = NA,
  return_plot_table = c("both", "plot only", "table only")
)
}
\arguments{
\item{df}{master_tb dataset}

\item{drop_years}{years to drop e.g. c(2022) or c(2021, 2022)}

\item{return_plot_table}{do you want to return a plot, table version or both.}
}
\value{
returns a table or visually plot of the number of positive bTB animals
}
\description{
\code{top} Extracts positive bTB cases per year by skin test,
lab and GIF.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #cases <- all_cases_per_year(master_tb)
 #cases <- all_cases_per_year(master_tb, drop_years = c(2022))
 #cases <- all_cases_per_year(master_tb, drop_years = c(2022), return_plot_table = "plot only")
 #cases <- all_cases_per_year(master_tb, drop_years = c(2022), return_plot_table = "table only")
 }
}
}
