% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd_dataset_fun.R
\name{bd_dataset_fun}
\alias{bd_dataset_fun}
\title{bd_dataset_fun}
\usage{
bd_dataset_fun(df)
}
\arguments{
\item{df}{Master bTB dataset (master_tb).}
}
\value{
returns BD dataset which only contains information on individual
BDs. If a herd never had a BD, they will not be in this file. Each row is
a summary of the BD (e.g. number of reactors over the BD period etc.). A
herd can have multiple rows if it has had more than one BD.
}
\description{
\code{top} Creates BD file containing information on each BD.
}
\details{
DETAILS
}
\examples{
\dontrun{
#bd_df <- bd_dataset_fun(master_tb)
}
}
