% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd_df_cvera_read_in.R
\name{bd_df_cvera_read_in}
\alias{bd_df_cvera_read_in}
\title{Read the Latest bTB breakdown dataset (bd_df)}
\usage{
bd_df_cvera_read_in(base_path = "N:/data/tb/master_tb_data_jamie_m")
}
\arguments{
\item{base_path}{Character. Path containing folders of the form
\code{year_XXXX_qX}.}
}
\value{
A dataset read from the most recent \code{bd_df*.rds} file.
}
\description{
This function finds the most recent \code{year_XXXX_qX} folder inside a base
directory, identifies \code{.rds} files whose filename contains \code{"bd_df"},
and loads the most recently modified file.
}
\examples{
\dontrun{
bd_df_cvera_read_in("N:/data/tb/master_tb_data_jamie_m")
}

}
