% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd_during.R
\name{bd_during}
\alias{bd_during}
\title{bd_during}
\usage{
bd_during(df, day_to_check, format = "\%Y-\%m-\%d")
}
\arguments{
\item{df}{BD dataset (bd_df). Created using bd_dataset_fun (bd_df <- bd_dataset_fun(master_tb))}

\item{day_to_check}{date to check if herd was in a BD during this time}

\item{format}{date format, Default: '\%Y-\%m-\%d'}
}
\value{
returns new column in BD dataset.
}
\description{
\code{top} Check what herds were in BD on a particular date.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #bd_check <- bd_during(bd_df, "2017-02-27")
 }
}
}
