% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd_during_year.R
\name{bd_during_year}
\alias{bd_during_year}
\title{bd_during_year}
\usage{
bd_during_year(df, years_to_check)
}
\arguments{
\item{df}{BD dataset}

\item{years_to_check}{years to check if herd was in a BD during this time.}
}
\value{
returns new column in BD dataset.
}
\description{
\code{top} is a small function to not just present the first rows
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
newdf <- bd_during_year(bd_test, years_to_check = c(2005))
newdf <- bd_during_year(bd_test, years_to_check = c(2005:2006)) }
}
}
