% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd_within_time_period.R
\name{bd_within_time_period}
\alias{bd_within_time_period}
\title{bd_within_time_period}
\usage{
bd_within_time_period(df, year_bd_started, time_period)
}
\arguments{
\item{df}{BD dataset (bd_df). Created using bd_dataset_fun (bd_df <- bd_dataset_fun(master_tb))}

\item{year_bd_started}{years to check if herd was in a BD during this time.}

\item{time_period}{Time period (in days) to see if herd was in BD prior to BD of interest which started in (year_bd_started).}
}
\value{
returns new indicator column(s) in BD dataset (0 = not in BD within time_period, 1 = BD within time_period).
}
\description{
\code{top}  Check what herds were in BD in a particular year
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
bd_df <- bd_within_time_period(bd_df, 2015, 365)
bd_df <- bd_within_time_period(bd_df, 2016, 730)}
}
}
