% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_movement_db.R
\name{connect_to_movement_db}
\alias{connect_to_movement_db}
\title{Connect to the Animal Identification and Movement (AIM) SQL Database (db)}
\usage{
connect_to_movement_db(database_name = final1)
}
\arguments{
\item{database_name}{Name of the database. Default is "final1", this may change depending on Jamie T. "network" is also available.}
}
\value{
A DBI connection object.
}
\description{
Establishes a connection in R to the movement database on the server. Note,
you need to have set up a connection with the server via your PC initially,
with correct password etc (once-off set-up).
}
\examples{
\dontrun{
con_move <- connect_to_movement_db(final1)
}

}
