% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_vars.R
\name{core_vars}
\alias{core_vars}
\title{core_vars}
\usage{
core_vars(df, ...)
}
\arguments{
\item{df}{Master bTB dataset (master_tb). (...) = add in additional variables}
}
\value{
returns dataset with core variables
}
\description{
\code{top} Simple function to extract (e.g. dplyr::select)
core variables from master_tb dataset, useful for quick sanity checks and can
be added into a pipe. Saves having to type them out manually.
}
\details{
DETAILS
}
\examples{
\dontrun{
#master_tb \%>\%
# filter(total_reactor_skin > 10) \%>\%
# core_vars()
#add variables
#master_tb \%>\%
# filter(total_reactor_skin > 10) \%>\%
# core_vars(county)
#remove one of the core variables
#master_tb \%>\%
# filter(total_reactor_skin > 10) \%>\%
# core_vars(-test_type)
}
}
