% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herd_plot.R
\name{herd_plot}
\alias{herd_plot}
\title{herd_plot}
\usage{
herd_plot(
  herd_no_character,
  start_date = NULL,
  end_date = NULL,
  format = "\%Y-\%m-\%d",
  include_tables = FALSE,
  alternative_herd_no_name = NULL
)
}
\arguments{
\item{herd_no_character}{herd number}

\item{start_date}{start date of plot search. Format, '\%Y-\%m-\%d'. Default is NULL so all years are plotted.}

\item{end_date}{end date of plot search. Format, '\%Y-\%m-\%d'. Default is NULL so all years are plotted.}

\item{format}{date format, Default: '\%Y-\%m-\%d'}

\item{include_tables}{Include herd-level tables of master_tb and breakdown datatset bd_df for the herd of interest. Default is FALSE which produces interactive plot without tables.}

\item{alternative_herd_no_name}{If you want to change the herd_no in title to something else.}
}
\value{
returns new column in BD dataset.
}
\description{
\code{top} is a small function to not just present the first rows
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
herd_plot("x1234567")
herd_plot("x1234567", start_date = "2010-01-01")
herd_plot("x1234567", include_tables = TRUE, alternative_herd_no_name = "other name/number")

 }
}
}
