% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_cvera_snippets.R
\name{install_cvera_snippets}
\alias{install_cvera_snippets}
\title{Install RStudio cvera snippets for cvera package}
\usage{
install_cvera_snippets()
}
\value{
Invisible NULL
}
\description{
Copies the package-provided R snippets into the user's RStudio snippets file.
If a previous version of the snippets block exists, it will be replaced.
}
\examples{
\dontrun{
if(interactive()){
install_cvera_snippets()
 }
}
}
