% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_tb_cvera_read_in.R
\name{master_tb_cvera_read_in}
\alias{master_tb_cvera_read_in}
\title{Automatically read in the latest available version of master_tb Dataset}
\usage{
master_tb_cvera_read_in(base_path = "N:/data/tb/master_tb_data_jamie_m")
}
\arguments{
\item{base_path}{Character. Path containing folders of the form
\code{year_XXXX_qX}.}
}
\value{
A dataset read from the most recent \code{master_tb*.rds} file.
}
\description{
This function finds the most recent \code{year_XXXX_qX} folder inside a base
directory, identifies \code{.rds} files whose filename contains \code{"master_tb"},
and loads the most recently modified file.
}
\examples{
\dontrun{
master_tb_read_in("N:/data/tb/master_tb_data_jamie_m")
}

}
