% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_types_sictt_table.R
\name{test_types_sictt_table}
\alias{test_types_sictt_table}
\title{View skin test type definitions (SICTT)}
\usage{
test_types_sictt_table(test_type_filter = NULL)
}
\arguments{
\item{test_type_filter}{Optional character vector of test types to filter.
If NULL (default), all rows are returned.}
}
\value{
A gt table.

#' @examples
\dontrun{
# print all sictt definitions:
test_types_sictt_table()
# only print specific ones:
}
}
\description{
Reads the test types file definitions and returns a formatted table.
}
