% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd_during.R
\name{bd_during}
\alias{bd_during}
\title{bd_during}
\usage{
bd_during(df, day_to_check, format = "\%Y-\%m-\%d")
}
\arguments{
\item{df}{BD dataset}

\item{day_to_check}{date to check if herd was in a BD during this time}

\item{format}{date format, Default: '\%Y-\%m-\%d'}
}
\value{
returns new column in BD dataset.
}
\description{
\code{top} is a small function to not just present the first rows
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #bd_test <- bd_during(bd_test, "2017-02-27")
 }
}
}
