% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd_during_year.R
\name{bd_during_year}
\alias{bd_during_year}
\title{bd_during_year}
\usage{
bd_during_year(df, years_to_check)
}
\arguments{
\item{df}{BD dataset (bd_df). Created using bd_dataset_fun (bd_df <- bd_dataset_fun(master_tb))}

\item{years_to_check}{years to check if herd was in a BD during this time.}
}
\value{
returns new indicator column(s) in BD dataset (0 = not in BD, 1 = in BD).
}
\description{
\code{top}  Check what herds were in BD in a particular year
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
newdf <- bd_during_year(bd_df, years_to_check = c(2005))
newdf <- bd_during_year(bd_df, years_to_check = c(2005:2006)) }
}
}
