% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herd_plot.R
\name{herd_plot}
\alias{herd_plot}
\title{herd_plot}
\usage{
herd_plot(
  df,
  herd_no_character,
  start_date = NULL,
  end_date = NULL,
  format = "\%Y-\%m-\%d",
  plotly_output = TRUE
)
}
\arguments{
\item{df}{master_tb dataset}

\item{herd_no_character}{herd number}

\item{start_date}{start date of plot search. Format, '\%Y-\%m-\%d'. Default is NULL so all years are plotted.}

\item{end_date}{end date of plot search. Format, '\%Y-\%m-\%d'. Default is NULL so all years are plotted.}

\item{format}{date format, Default: '\%Y-\%m-\%d'}

\item{plotly_output}{Decide whether plot output is static or interactive. Default is TRUE which returns interactive plotly graph.}
}
\value{
returns new column in BD dataset.
}
\description{
\code{top} is a small function to not just present the first rows
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
p <- herd_plot(master_tb, "x1234567")
p
p <- herd_plot(master_tb, "x1234567", start_date = "2010-01-01")
p
p <- herd_plot(master_tb, "x1234567", plotly_output = FALSE) #static plot
p
 }
}
}
